extracting/requirements.txt
psycopg2-binary
python-dotenv
boto3
pandas
moto
pytest